minor logging, p2p work in progress, eth

+++ src/Nethermind/Nethermind.Network/P2P/NettyP2PHandler.cs
@@ -25,7 +25,7 @@ namespace Nethermind.Network.P2P
 {
     public class NettyP2PHandler : SimpleChannelInboundHandler<Packet>
     {
-        public static int Version = 5; // TODO: move somewhere else
+        public static byte Version = 5; // TODO: move somewhere else
         
         private readonly ILogger _logger;
         private readonly IMessageSerializationService _serializationService;
